requirements.txt
# Python runtime dependencies for T28 IO Ring Orchestrator
# Python 3.7+

# Required by scripts/check_virtuoso_connection.py and assets/utils/bridge_utils.py
python-dotenv>=1.0,<2.0

# Required by assets/core/layout/layout_visualizer.py
matplotlib>=3.5,<4.0

# Optional dependency when USE_RAMIC_BRIDGE is false and skillbridge mode is used.
# Install manually if your flow uses skillbridge:
#   pip install skillbridge
